Remove cloudfront URL's.
I've consistently found that using CloudFront on top of the raw S3 bucket
has *worse* performance than just downloading from S3 directly. This might
be because we tend to hit cold CloudFront caches since there are not
e.g. thousands of downloads of this per day. I've elected to revert
using CloudFront for now and we can switch it back on if we want
it.