Update for fully packaged Shark 0.8.0 and 0.8.1.
Previously the Shark 0.8.0 package was based on a SNAPSHOT
release and not using the finalized Shark packaging format.

There was a minor packaging change in that Hive is no longer
included in the binary Shark releases. So that is downloaded
seperately. Some care is taken to make sure that it still
works with older versions of Shark.